ALL ON main.py
Added title() function, this function lets the end-user edit the title when creating the window and after the window is created.
Fixed transparent fucntion so the scrollbar and listbox to hide,
Fixed the reopen function to add back the listbox and scroll bar to correctly fill the y-axis and anchor to the left side of the Tk window.
Changed the .pack_configure to .place for men_collap and men_open (Also moved men_open button .place to the bottom of the .py file)
Moved the men_collap and men_open button in the Tk window to the top right side of the screen.
Added the from tkinter.messagebox import showinfo for the title function.
Added the from tkinter.simpledialog import askstring for inputting title into the messsage box.
Moved box = tk.Tk() to the top of the main.py file
Removed .destroy() from the functions
Changed the width and height of the text_box.